refact(FogMaze) : Fog maze puzzle script들의 책임 분배
